requirements.txt
Django #==1.5.1
South #==0.8.1
argparse==1.2.1
distribute>=0.6.24
django-tables2>=0.13.0
django-tastypie #==0.9.15
docutils==0.10
ipython==0.13.1
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==2.1
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
